Separate CScriptCompressor
